requirements.txt
absl-py==2.1.0
Adafruit-Blinka==8.56.0
adafruit-circuitpython-busdevice==5.2.11
adafruit-circuitpython-connectionmanager==3.1.3
adafruit-circuitpython-requests==4.1.10
adafruit-circuitpython-typing==1.11.2
adafruit-circuitpython-vl6180x==1.4.16
Adafruit-PlatformDetect==3.77.0
Adafruit-PureIO==1.1.11
attrs==25.3.0
binho-host-adapter==0.1.6
blinker==1.9.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
contourpy==1.3.0
cycler==0.12.1
dotenv==0.9.9
Flask==3.1.0
flask-cors==5.0.1
flatbuffers==25.2.10
fonttools==4.56.0
idna==3.10
importlib_metadata==8.6.1
importlib_resources==6.5.2
itsdangerous==2.2.0
jax==0.4.30
jaxlib==0.4.30
Jinja2==3.1.6
kiwisolver==1.4.7
MarkupSafe==3.0.2
matplotlib==3.9.4
mediapipe==0.10.21
ml_dtypes==0.5.1
numpy==1.26.4
opencv-contrib-python==4.11.0.86
opencv-python==4.11.0.86
opt_einsum==3.4.0
packaging==24.2
paho-mqtt==2.1.0
pillow==11.1.0
protobuf==4.25.6
pycparser==2.22
pyftdi==0.56.0
pygame==2.6.1
pyparsing==3.2.1
pyserial==3.5
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pyusb==1.3.1
qrcode==8.0
requests==2.32.3
scipy==1.13.1
sentencepiece==0.2.0
six==1.17.0
sounddevice==0.5.1
toml==0.10.2
typing_extensions==4.13.2
urllib3==2.3.0
Werkzeug==3.1.3
zipp==3.21.0
